requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cftime==1.6.4.post1
charset-normalizer==3.4.3
Django==5.2
django-environ==0.12.0
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
netCDF4==1.7.2
numpy==2.3.1
packaging==25.0
pandas==2.3.1
phonenumberslite==9.0.12
pillow==11.3.0
psycopg2==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
xarray==2025.7.1
